requirements.txt
Flask
requests
python-dotenv
redis
pytest
